regress: move mocked io to its own source file
No functional change.
